% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing.R
\name{bulkSample}
\alias{bulkSample}
\title{Quantify the spatial distribution of mutants}
\usage{
bulkSample(tumor, pos, cube_length = 5, threshold = 0.05)
}
\arguments{
\item{tumor}{The output of \link[TumorGenerator]{simulateTumor}.}

\item{pos}{TODO}

\item{cube_length}{TODO}

\item{threshold}{TODO}
}
\description{
Quantify the spatial distribution of mutants
}
