% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTumor.R
\name{generateTumor}
\alias{generateTumor}
\title{Spatial simulation of tumor growth}
\usage{
generateTumor(
  N = 1e+05,
  b = 0.25,
  d = 0.13,
  u = 0.01,
  du = 0.003,
  s = 1.1,
  cutoff = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{N}{Number of cells in the tumor}

\item{b}{Cell division rate}

\item{d}{Cell death rate}

\item{u}{Mutation rate. When a cell divides, both daughter cell acquire $Pois(u)$ genetic alterations}

\item{du}{The probability that a genetic alteration is a driver mutation}

\item{s}{The selective advantage conferred to a driver mutation. A cell with k
driver mutations is given birth rate $bs^k$.}
}
\value{
A list \code{out}.
}
\description{
Simulate the spatial growth of a tumor using a multi-type branching
process on the three-dimensional integer lattice.
}
