% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotSlice}
\alias{plotSlice}
\title{2D cross section of the simulated tumor}
\usage{
plotSlice(tumor, slice.dim = "x", level = 0, plot.type = "normal")
}
\arguments{
\item{tumor}{A list which is the output of \code{\link{simulateTumor}()}.}

\item{slice.dim}{One of "x", "y" or "z", which denotes the dimension which will be fixed to obtain a 2D cross section.}

\item{level}{Which value will the dimension given in \code{slice.dim} be fixed at?}

\item{plot.type}{Which type of plot to draw. "Normal" assigns a random rgb value to each allele while
"heat" colors cells with more mutations red and cells with fewer mutations blue. This is exactly the same as \code{plot.type}
in \code{visualizeTumor}.}
}
\value{
None.
}
\description{
2D cross section of the simulated tumor.
}
\author{
Phillip B. Nicol
}
