% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progDAG.R
\name{progressionChain}
\alias{progressionChain}
\title{Create a linear chain graph to describe the order of mutations}
\usage{
progressionChain(n)
}
\arguments{
\item{n}{Number of vertices in the chain}
}
\value{
A matrix with 4 columns and n-1 rows which will be accepted as input to
\code{\link{simulateTumor}()}.
}
\description{
A helper function for \code{\link{simulateTumor}()} which returns to the user
the edge list for a linear chain.
}
\examples{
G <- progressionChain(3)
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
