% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progDAG.R
\name{progressionDAG_from_igraph}
\alias{progressionDAG_from_igraph}
\title{Define the progression of mutations from an \code{igraph} object}
\usage{
progressionDAG_from_igraph(iG)
}
\arguments{
\item{iG}{An igraph object for a directed acyclic graph.}
}
\value{
A matrix with 4 columns which contains the edges of the graph as well as the rate of 
crossing each edge and the selective advantage/disadvantage obtained by crossing each edge.
}
\description{
A helper function for \code{\link{simulateTumorMTBP}()} which returns to the user
the edge list for a DAG which is defined as an \code{igraph} object.
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
