% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing.R
\name{randomBulkSamples}
\alias{randomBulkSamples}
\title{Simulate bulk sequencing data}
\usage{
randomBulkSamples(tumor, nsamples, cube_length = 5, threshold = 0.05)
}
\arguments{
\item{tumor}{A list which is the output of \code{\link{simulateTumor}}.}

\item{nsamples}{The number of samples to take.}

\item{cube_length}{The side length of the cube in}

\item{threshold}{TODO}
}
\description{
Simulate bulk sequencing data by takign a local sample of the tumor
and computing the variant allele frequencies of the various mutations.
}
