% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing.R
\name{singleCell}
\alias{singleCell}
\title{Simulate single cell sequencing data}
\usage{
singleCell(tumor, pos, noise = 0)
}
\arguments{
\item{tumor}{A list which is the output of \code{\link{simulateTumor}}.}

\item{pos}{A vector of length 3 giving the (x,y,z) coordinates of the cell to sample.}

\item{noise}{The false negative rate.}
}
\description{
Simulate single cell sequencing data by selecting a cell at a specified position
}
\details{
TODO
}
\references{

}
\author{
Phillip B. Nicol
}
