% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatialDistribution}
\alias{spatialDistribution}
\title{Quantify the spatial distribution of mutants}
\usage{
spatialDistribution(tumor, N = 500, cutoff = 0.01, make.plot = TRUE)
}
\arguments{
\item{tumor}{The output of \link[TumorGenerator]{simulateTumor}.}

\item{N}{TODO}

\item{cutoff}{TODO}

\item{make.plot}{TODO}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Quantify the spatial distribution of mutants
}
