% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualizeTumor}
\alias{visualizeTumor}
\title{Interactive visualization of the simulated tumor}
\usage{
visualizeTumor(tumor, plot.type = "normal", background = "black", axes = FALSE)
}
\arguments{
\item{tumor}{A list which is the output of \code{\link{simulateTumor}()}.}

\item{plot.type}{Which type of plot to draw. "Normal" assigns a random rgb value to each genotype while
"heat" colors cells with more mutations red and cells with fewer mutations blue.}

\item{background}{If rgl is installed, this will set the color of the background}

\item{axes}{Will include axes (rgl only).}
}
\value{
None.
}
\description{
Interactive visualization of the simulated tumor using the \code{rgl} package (if available).
}
\details{
If \pkg{rgl} is installed, then the plots will be interactive. If \pkg{rgl} is unavailable, static plots will
be created with \pkg{scatterplot3d}. Since plotting performance with \pkg{scatterplot3d} is reduced, it is strongly
recommended that \pkg{rgl} is installed for optimal use of this function.
}
\author{
Phillip B. Nicol
}
